corrected flaw in management of vertical linearizations
UpFiLmb1.back() == INF used to mean "a vertical linearization has been found" but this is no longer true, hence the information had to be encoded in MPchgs (now a char rather than a bool). This does not erroneously force the t-changing phase of a NS to be skipped if not all components have a UB computed.