fixed several blunders related to upper al lower cutoffs
At the very least, a std::abs() was missing when using m1 (which can be < 0). Most importantly, std::min and std::max were inverted when computing the second part of the upper and lower cutoff, which could lead to UpCutOff < LwCutOff happening.
Also, finite UpCutOff and LwCutOff are now only imposed if both the SS and NS condition do *not* hold already.