improved BundleSolverState
Now all the information about the function values (lw and up) in Lambda are saved, so that upon restore the algorithm can restart precisely from where it left. Previously one had to perform a SS to whatever the optimal solution of the master problem was, which may be a much worse point than the stability centre.